chore: rename crowpanel_101 to crowpanel to cover all range of screen sizes
