+++ .github/workflows/test-all-builds.yml
@@ -11,7 +11,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        board: [wave_4b, wave_35, wave_5, wave_43, crowpanel_101]
+        board: [wave_4b, wave_35, wave_5, wave_43, crowpanel]
     name: build (${{ matrix.board }})
     steps:
     - name: Checkout repo
@@ -81,11 +81,11 @@ jobs:
       with:
         name: firmware-wave_43
         path: flasher/firmware/wave_43
-    - name: Download crowpanel_101 firmware
+    - name: Download crowpanel firmware
       uses: actions/download-artifact@v4.1.3
       with:
-        name: firmware-crowpanel_101
-        path: flasher/firmware/crowpanel_101
+        name: firmware-crowpanel
+        path: flasher/firmware/crowpanel
     - name: Stage branding assets
       run: cp -r branding flasher/branding
     - name: Upload Pages artifact
