+++ .gitignore
@@ -4,7 +4,7 @@ build_wave_35/
 build_wave_4b/
 build_wave_43/
 build_wave_5/
-build_crowpanel_101/
+build_crowpanel/
 sdkconfig
 sdkconfig.old
 managed_components/
