+++ components/bsp_common/CMakeLists.txt
@@ -3,7 +3,7 @@
 
 set(srcs "")
 # Boards without a dedicated PMIC chip get the no-op stub.
-if(NOT CONFIG_KERN_BOARD_WAVE_35 AND NOT CONFIG_KERN_BOARD_CROWPANEL_101)
+if(NOT CONFIG_KERN_BOARD_WAVE_35 AND NOT CONFIG_KERN_BOARD_CROWPANEL)
     list(APPEND srcs "pmic_stub.c")
 endif()
