+++ components/crowpanel/include/bsp/esp-bsp.h
@@ -1,0 +1,2 @@
+#pragma once
+#include "bsp/crowpanel.h"
