--- components/crowpanel_101/include/bsp/esp-bsp.h
@@ -1,2 +1,0 @@
-#pragma once
-#include "bsp/crowpanel_101.h"
