+++ flasher/index.html
@@ -178,7 +178,7 @@
         <option value="wave_35">Waveshare ESP32-P4 3.5&quot; (320×480 SPI)</option>
         <option value="wave_5">Waveshare ESP32-P4 5&quot; (720×1280 MIPI DSI)</option>
         <option value="wave_43">Waveshare ESP32-P4 4.3&quot; (480×800 MIPI DSI)</option>
-        <option value="crowpanel_101">CrowPanel Advanced 10.1&quot; ESP32-P4 (1024×600 MIPI DSI)</option>
+        <option value="crowpanel">CrowPanel Advanced ESP32-P4 7&quot; / 10.1&quot; (1024×600 MIPI DSI)</option>
       </select>
       <div class="file-list" id="ciFileList"></div>
     </div>
