+++ scripts/format.sh
@@ -26,7 +26,7 @@ DIRS=(
     "$REPO_ROOT/components/wave_4b"
     "$REPO_ROOT/components/wave_35"
     "$REPO_ROOT/components/wave_43"
-    "$REPO_ROOT/components/crowpanel_101"
+    "$REPO_ROOT/components/crowpanel"
 )
 
 if $CHECK_MODE; then
