+++ sdkconfig.defaults.crowpanel
@@ -1,5 +1,5 @@
-# Board: CrowPanel Advanced 10.1" ESP32-P4 (1024x600 MIPI DSI)
-CONFIG_KERN_BOARD_CROWPANEL_101=y
+# Board: CrowPanel Advanced ESP32-P4 (7" / 10.1", 1024x600 MIPI DSI)
+CONFIG_KERN_BOARD_CROWPANEL=y
 CONFIG_ESPTOOLPY_FLASHMODE_QIO=y
 CONFIG_SD_BUS_WIDTH=1
 CONFIG_SD_INTERNAL_PULLUP=y
