+++ simulator/CMakeLists.txt
@@ -373,7 +373,7 @@ target_compile_definitions(kern_simulator PRIVATE
     $<$<STREQUAL:${SIM_BOARD},wave_35>:CONFIG_KERN_BOARD_WAVE_35=1>
     $<$<STREQUAL:${SIM_BOARD},wave_5>:CONFIG_KERN_BOARD_WAVE_5=1>
     $<$<STREQUAL:${SIM_BOARD},wave_43>:CONFIG_KERN_BOARD_WAVE_43=1>
-    $<$<STREQUAL:${SIM_BOARD},crowpanel_101>:CONFIG_KERN_BOARD_CROWPANEL_101=1>
+    $<$<STREQUAL:${SIM_BOARD},crowpanel>:CONFIG_KERN_BOARD_CROWPANEL=1>
 )
 
 target_compile_options(kern_simulator PRIVATE
